+++ Content/Items/Reds/SpareDroneParts.cs
@@ -7,7 +7,16 @@
 
         public override string PickupText => "Your drones fire faster, have less cooldowns, shoot missiles, and gain a bonus chaingun.";
 
-        public override string DescText => "Gain <style=cIsDamage>Col. Droneman.</style> Drones gain <style=cIsDamage>+50%</style> <style=cStack>(+50% per stack)</style> attack speed and cooldown reduction. Drones gain <style=cIsDamage>10%</style> chance to fire a <style=cIsDamage>missile</style> on hit, dealing <style=cIsDamage>300%</style> TOTAL damage. Drones gain an <style=cIsDamage>automatic chain gun</style> that deals <style=cIsDamage>6x30% damage</style>.";
+        public override string DescText => "Gain <style=cIsDamage>Col. Droneman.</style> Drones gain <style=cIsDamage>+" + d(attackSpeedCdr) + "</style> <style=cStack>(+" + d(attackSpeedCdr) + " per stack)</style> attack speed and cooldown reduction, a <style=cIsDamage>10%</style> chance to fire a <style=cIsDamage>missile</style> on hit, dealing <style=cIsDamage>300%</style> TOTAL damage and an <style=cIsDamage>automatic chain gun</style> that deals <style=cIsDamage>6x" + d(chaingunDamage) + " damage</style>" + (chaingunBounces > 0 ? " and <style=cIsDamage>bounces</style> up to " + chaingunBounces + " times." : ".");
+
+        [ConfigField("Chaingun Bounces", 0)]
+        public static int chaingunBounces;
+
+        [ConfigField("Chaingun Damage", "Decimal.", 0.3f)]
+        public static float chaingunDamage;
+
+        [ConfigField("Attack Speed and Cooldown Reduction Gain", "Decimal.", 0.4f)]
+        public static float attackSpeedCdr;
 
         public override void Init()
         {
@@ -17,12 +26,48 @@
         public override void Hooks()
         {
             On.EntityStates.DroneWeaponsChainGun.FireChainGun.OnEnter += ChaingunChanges;
+            IL.RoR2.CharacterBody.RecalculateStats += CharacterBody_RecalculateStats;
+        }
+
+        private void CharacterBody_RecalculateStats(ILContext il)
+        {
+            ILCursor c = new(il);
+
+            if (c.TryGotoNext(MoveType.Before,
+                x => x.MatchLdloc(46),
+                x => x.MatchConvR4(),
+                x => x.MatchLdcR4(0.5f),
+                x => x.MatchMul()))
+            {
+                c.Index += 2;
+                c.Next.Operand = attackSpeedCdr;
+            }
+            else
+            {
+                Logger.LogError("Failed to apply Spare Drone Parts Attack Speed hook");
+            }
+
+            c.Index = 0;
+
+            if (c.TryGotoNext(MoveType.Before,
+                    x => x.MatchLdloc(87),
+                    x => x.MatchLdcR4(0.5f),
+                    x => x.MatchMul(),
+                    x => x.MatchStloc(87)))
+            {
+                c.Index += 1;
+                c.Next.Operand = 1f - attackSpeedCdr;
+            }
+            else
+            {
+                Logger.LogError("Failed to apply Spare Drone Parts Cooldown Reduction hook");
+            }
         }
 
         public static void ChaingunChanges(On.EntityStates.DroneWeaponsChainGun.FireChainGun.orig_OnEnter orig, EntityStates.DroneWeaponsChainGun.FireChainGun self)
         {
-            self.additionalBounces = 0;
-            self.damageCoefficient = 0.3f;
+            self.additionalBounces = chaingunBounces;
+            self.damageCoefficient = chaingunDamage;
             orig(self);
         }
     }